Move breadcrumbs and nav bar to `org.rust.ide.structure`
