+++ src/main/kotlin/org/rust/ide/structure/RsBreadcrumbsInfoProvider.kt
@@ -3,7 +3,7 @@
  * found in the LICENSE file.
  */
 
-package org.rust.ide.miscExtensions
+package org.rust.ide.structure
 
 import com.intellij.openapi.project.DumbService
 import com.intellij.psi.PsiElement
