+++ src/main/kotlin/org/rust/ide/structure/RsStructureViewModel.kt
@@ -20,7 +20,6 @@ import com.intellij.pom.Navigatable
 import com.intellij.ui.icons.RowIcon
 import com.intellij.util.PlatformIcons
 import com.intellij.util.containers.map2Array
-import org.rust.ide.miscExtensions.RsBreadcrumbsInfoProvider
 import org.rust.ide.presentation.getPresentationForStructure
 import org.rust.lang.core.psi.*
 import org.rust.lang.core.psi.ext.*
